hohealth-python/requirements.txt
pdf2image>=1.16.0
Pillow>=9.0.1
pytesseract>=0.3.9
python-poppler>=0.2.2